MEDIUM: mux-h2/trace: add tracing support for headers
Now we can make use of TRACE_PRINTF() to iterate over headers as they
are received or dumped. It's worth noting that the dumps may occasionally
be interrupted due to a buffer full or a realign, but in this case it
will be visible because the trace will restart from the first one. All
these headers (and trailers) may be interleaved with other connections'
so they're all preceeded by the pointer to the connectionn to help
discriminating them. The header number is indicated as well, however
they might not always be contiguous because some headers are dropped.
Others are built on the fly (status, method etc) and are not part of
the numbering scheme so they're numbered as "--".

Example of what an end-to-end H2 request gives:

  [00|h2|5|mux_h2.c:4753] h2c_decode_headers(): h2c=0x2ff46b0(F,FRP) hdr[00] l=1    :path: /
  [00|h2|5|mux_h2.c:4753] h2c_decode_headers(): h2c=0x2ff46b0(F,FRP) hdr[01] l=4    :scheme: http
  [00|h2|5|mux_h2.c:4753] h2c_decode_headers(): h2c=0x2ff46b0(F,FRP) hdr[02] l=15   :authority: 127.0.0.1:14446
  [00|h2|5|mux_h2.c:4753] h2c_decode_headers(): h2c=0x2ff46b0(F,FRP) hdr[03] l=3    :method: GET
  [00|h2|5|mux_h2.c:4753] h2c_decode_headers(): h2c=0x2ff46b0(F,FRP) hdr[04] l=25   user-agent: h2load nghttp2/1.50.0-DEV
  [00|h2|5|mux_h2.c:5447] h2s_bck_make_req_headers(): h2c=0x3002760(B,FRH) hdr[--] l=3    :method: GET
  [00|h2|5|mux_h2.c:5546] h2s_bck_make_req_headers(): h2c=0x3002760(B,FRH) hdr[--] l=4    :scheme: http
  [00|h2|5|mux_h2.c:5563] h2s_bck_make_req_headers(): h2c=0x3002760(B,FRH) hdr[--] l=15   :authority: 127.0.0.1:14446
  [00|h2|5|mux_h2.c:5590] h2s_bck_make_req_headers(): h2c=0x3002760(B,FRH) hdr[--] l=1    :path: /
  [00|h2|5|mux_h2.c:5674] h2s_bck_make_req_headers(): h2c=0x3002760(B,FRH) hdr[01] l=25   user-agent: h2load nghttp2/1.50.0-DEV
  [00|h2|5|mux_h2.c:4753] h2c_decode_headers(): h2c=0x3002760(B,FRP) hdr[00] l=3    :status: 200
  [00|h2|5|mux_h2.c:4753] h2c_decode_headers(): h2c=0x3002760(B,FRP) hdr[01] l=1    content-length: 0
  [00|h2|5|mux_h2.c:4753] h2c_decode_headers(): h2c=0x3002760(B,FRP) hdr[02] l=18   x-req: size=80, time=0 ms
  [00|h2|5|mux_h2.c:4753] h2c_decode_headers(): h2c=0x3002760(B,FRP) hdr[03] l=55   x-rsp: id=dummy, code=200, cache=1, size=0, time=0 ms (0 real)
  [00|h2|5|mux_h2.c:5145] h2s_frt_make_resp_headers(): h2c=0x2ff46b0(F,FRH) hdr[--] l=3    :status: 200
  [00|h2|5|mux_h2.c:5184] h2s_frt_make_resp_headers(): h2c=0x2ff46b0(F,FRH) hdr[00] l=1    content-length: 0
  [00|h2|5|mux_h2.c:5184] h2s_frt_make_resp_headers(): h2c=0x2ff46b0(F,FRH) hdr[01] l=18   x-req: size=80, time=0 ms
  [00|h2|5|mux_h2.c:5184] h2s_frt_make_resp_headers(): h2c=0x2ff46b0(F,FRH) hdr[02] l=55   x-rsp: id=dummy, code=200, cache=1, size=0, time=0 ms (0 real)

Note that for responses and possibly for backend requests, having the stream
ID would be desirable. Also, at some point the frontend/backend names would
be useful but that's a more general problem.