CLEANUP: trace: use __func__ instead of __FUNCTION__
Among the expected facilities awaited for the TRACE_PRINTF(), one desirable
feature will be to centralize some trace formatting functions (e.g. produce
header lists) with the name of the calling function being reporterd and
not of the dumping function. It was verified that this can be achieved
by #defining __func__ to temporarily point to a local argument around a
call to a trace macro, but it's less sure that doing so with the old gcc
specific __FUNCTION__ will be as portable. So let's first update the
macros to use the more portable __func__.