MEDIUM: listener/config: make the "thread" parser rely on thread_sets
Instead of reading and storing a single group and a single mask for a
"thread" directive on a bind line, we now store the complete range in
a thread set that's stored in the bind_conf. The bind_parse_thread()
function now just calls parse_thread_set() on an empty set, and
thread_resolve_group_mask() was updated to support retrieving thread
group numbers or absolute thread numbers directly from the pre-filled
thread_set, and continue to feed bind_tgroup and bind_thread. The CLI
parsers which were pre-initialized to set the bind_tgroup to 1 do not
need to do it anymore, instead they preset the whole group 1 in the
thread_set (once all of this is finished it should not even be
necessary anymore since the purpose was to save the user from having
to append a "thread" directive on the stats socket).

At this step it's already possible to start with all previous valid
configs as well as extended ones supporting comma-delimited thread
sets. In addition the parser already accepts large ranges spanning
multiple groups, but since the underlying listeners infrastructure
is not read, for now we're maintaining a specific check against this
at the higher level of the config validity check.

The patch is a bit large because thread resolution is performed in
multiple steps, so we need to adjust all of them at once to preserve
functional and technical consistency.